Summary: TSIA. This was a 2800 line file.
Relevant Issues: N/A

Type of change: /kind cleanup

Test Plan: Smoke/Unit tests. Nothing actually changes here; just some
constants and functions moving to new files.
